Update : 디렉토리 구조 정의 및 필요한 reference만 사용
